Fold whitespace tokens into a single flag on Token
Improves parsing ergonomics. Whitespace has to be preserved to figure out the
semantics of multiple-character symbols like <= but making it a separate
variant on `Token` means that whitespace padding has to be parsed everywhere.
This change folds whitespace padding into a flag on tokens that can be
easily ignored in all cases but still handled where it is significant.